+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add the account name to the account deletion-confirmation dialog
 - Add a hover tool-tip with account path and size when hovering over the email address of an account
   (this is useful when importing a backup of an account you already have and can't distinguish which is the old account and which is the imported one)
+- Add context menu to gallery
 
 ### Changed
 - Change "More info" translation to "Message Details"